Fixed installation and added test for it
I added some files to test whether `make install` works properly. This
has to be invoked manually (not as part of ctest) but can be done in
the CI. As part of this I caught some errors in how things were being
installed, which I fixed. I also changed how coverage reports are
generated, so they no longer depend on a particular job exiting
successfully and instead just build with whatever gcov data happens to
be available.